+++ packages/react-native-sdk/expo-config-plugin/src/index.ts
@@ -1,18 +1,35 @@
-import { ConfigPlugin, withPlugins } from '@expo/config-plugins';
-import withStreamVideoReactNativeSDKAppDelegate from './withAppDelegate';
+import {
+  ConfigPlugin,
+  createRunOncePlugin,
+  withPlugins,
+} from '@expo/config-plugins';
+import withStreamVideoReactNativeSDKAppDelegate from './withStreamVideoReactNativeSDKAppDelegate';
+import withPushAppDelegate from './withPushAppDelegate';
 import withStreamVideoReactNativeSDKMainApplication from './withMainApplication';
 import withStreamVideoReactNativeSDKAndroidPermissions from './withAndroidPermissions';
 import withStreamVideoReactNativeSDKManifest from './withAndroidManifest';
 import withStreamVideoReactNativeSDKiOSInfoPList from './withiOSInfoPlist';
+import { ConfigProps } from './common/types';
 
-const withStreamVideoReactNativeSDK: ConfigPlugin = (config) => {
+// path should be relative to dist
+const pkg = require('../../package.json');
+
+const withStreamVideoReactNativeSDK: ConfigPlugin<ConfigProps> = (
+  config,
+  props,
+) => {
   return withPlugins(config, [
+    () => withPushAppDelegate(config, props),
     withStreamVideoReactNativeSDKAppDelegate,
     withStreamVideoReactNativeSDKMainApplication,
     withStreamVideoReactNativeSDKAndroidPermissions,
     withStreamVideoReactNativeSDKManifest,
-    withStreamVideoReactNativeSDKiOSInfoPList,
+    () => withStreamVideoReactNativeSDKiOSInfoPList(config, props),
   ]);
 };
 
-export default withStreamVideoReactNativeSDK;
+export default createRunOncePlugin(
+  withStreamVideoReactNativeSDK,
+  pkg.name,
+  pkg.version,
+);