+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsGrid.tsx
@@ -4,30 +4,29 @@ import { useCallStateHooks } from '@stream-io/video-react-bindings';
 import { useDebouncedValue } from '../../../utils/hooks/useDebouncedValue';
 import {
   CallParticipantsList as DefaultCallParticipantsList,
-  CallParticipantsListProps,
   CallParticipantsListComponentProps,
 } from '../CallParticipantsList/CallParticipantsList';
 import { ComponentTestIds } from '../../../constants/TestIds';
 import { useTheme } from '../../../contexts/ThemeContext';
+import { CallContentProps } from '../CallContent';
+import { ParticipantViewComponentProps } from '../../Participant';
 
 /**
  * Props for the CallParticipantsGrid component.
  */
-export type CallParticipantsGridProps = CallParticipantsListComponentProps & {
-  /**
-   * Component to customize the CallParticipantsList.
-   */
-  CallParticipantsList?: React.ComponentType<CallParticipantsListProps> | null;
-  /**
-   * Boolean to decide if local participant will be visible in the grid when there is 1:1 call.
-   */
-  showLocalParticipant?: boolean;
-  /**
-   * Check if device is in landscape mode.
-   * This will apply the landscape mode styles to the component.
-   */
-  landscape?: boolean;
-};
+export type CallParticipantsGridProps = ParticipantViewComponentProps &
+  Pick<CallContentProps, 'supportedReactions' | 'CallParticipantsList'> &
+  Pick<CallParticipantsListComponentProps, 'ParticipantView'> & {
+    /**
+     * Boolean to decide if local participant will be visible in the grid when there is 1:1 call.
+     */
+    showLocalParticipant?: boolean;
+    /**
+     * Check if device is in landscape mode.
+     * This will apply the landscape mode styles to the component.
+     */
+    landscape?: boolean;
+  };
 
 /**
  * Component used to display the list of participants in a grid mode.
@@ -41,6 +40,7 @@ export const CallParticipantsGrid = ({
   ParticipantView,
   VideoRenderer,
   showLocalParticipant = false,
+  supportedReactions,
   landscape,
 }: CallParticipantsGridProps) => {
   const {
@@ -89,6 +89,7 @@ export const CallParticipantsGrid = ({
       {CallParticipantsList && (
         <CallParticipantsList
           participants={participants}
+          supportedReactions={supportedReactions}
           landscape={landscape}
           {...participantViewProps}
         />