+++ packages/react-native-sdk/src/utils/push/rxSubjects.ts
@@ -1,4 +1,13 @@
 import { BehaviorSubject } from 'rxjs';
+import { NonRingingPushEvent } from '../StreamVideoRN/types';
+
+/**
+ * This rxjs subject is used to store the call cid of the accepted incoming call from push notification
+ * Note: it is should be subscribed only when a user has connected to the websocket of Stream
+ */
+export const pushNonRingingCallData$ = new BehaviorSubject<
+  { cid: string; type: NonRingingPushEvent } | undefined
+>(undefined);
 
 /**
  * This rxjs subject is used to store the call cid of the accepted incoming call from push notification