+++ sample-apps/react/egress-composite/src/main.tsx
@@ -1,4 +1,6 @@
 import { createRoot } from 'react-dom/client';
+import * as Sentry from '@sentry/react';
+
 import {
   ConfigurationContext,
   ConfigurationValue,
@@ -7,9 +9,21 @@ import {
 import { CompositeApp } from './CompositeApp';
 
 import '@stream-io/video-react-sdk/dist/css/styles.css';
-
 // Uncomment this line to test your own custom CSS
 // import cssUrl from '../public/example/custom.css?url';
+
+if (import.meta.env.MODE === 'production') {
+  Sentry.init({
+    dsn: import.meta.env.VITE_EGRESS_SENTRY_DNS,
+    tracePropagationTargets: ['video-layout.getstream.io'],
+    integrations: [new Sentry.BrowserTracing(), new Sentry.Replay()],
+    tracesSampleRate: 0.9,
+    replaysSessionSampleRate: 0.1,
+    replaysOnErrorSampleRate: 1.0,
+    attachStacktrace: true,
+    maxBreadcrumbs: 10,
+  });
+}
 
 // @ts-expect-error TODO: this is a global function, we need to declare it
 window.setupLayout = (configuration: ConfigurationValue) => {